Add ability to manage adaptive LED

+++ lib/Pin/Pin.h
@@ -24,5 +24,5 @@ public:
 
   void writeLow();
 
-  void modulate(float dutyFactor);
+  void modulate(int dutyFactor);
 };
